chore: reformat logger
