--- src/bridge/logging/__init__.py
@@ -1,7 +1,0 @@
-"""
-Logging utilities for the bridge module.
-"""
-
-from .logging import setup_logging
-
-__all__ = ["setup_logging"]
